Address MCP lifecycle review comments
