+++ src/ipc/utils/mcp_shutdown.ts
@@ -1,4 +1,5 @@
 import { mcpManager } from "./mcp_manager";
+import { settleWithinTimeout } from "./promise_utils";
 
 const MCP_SHUTDOWN_TIMEOUT_MS = 2_000;
 
@@ -19,23 +20,6 @@ type McpBeforeQuitHandlerOptions = {
   cleanup?: () => Promise<void>;
   timeoutMs?: number;
 };
-
-function settleWithinTimeout(
-  promise: Promise<unknown>,
-  timeoutMs: number,
-): Promise<void> {
-  return new Promise((resolve) => {
-    let finished = false;
-    const finish = () => {
-      if (finished) return;
-      finished = true;
-      clearTimeout(timeout);
-      resolve();
-    };
-    const timeout = setTimeout(finish, timeoutMs);
-    void promise.then(finish, finish);
-  });
-}
 
 /**
  * Pause Electron's first before-quit event until MCP cleanup settles, then
