+++ src/ipc/utils/promise_utils.ts
@@ -1,0 +1,17 @@
+/** Resolve once a promise settles or the timeout expires, whichever comes first. */
+export function settleWithinTimeout(
+  promise: Promise<unknown>,
+  timeoutMs: number,
+): Promise<void> {
+  return new Promise((resolve) => {
+    let finished = false;
+    const finish = () => {
+      if (finished) return;
+      finished = true;
+      clearTimeout(timeout);
+      resolve();
+    };
+    const timeout = setTimeout(finish, timeoutMs);
+    void promise.then(finish, finish);
+  });
+}
